Update package versions and standardize size token naming for consistency, fixing json structure for color tokens.
