+++ wizards/space_wiz.js
@@ -366,7 +366,7 @@ const generateSpacingTokens = (
           "3xs",
           "2xs",
           "xs",
-          "s",
+          "sm",
           "md",
           "lg",
           "xl",
@@ -557,7 +557,7 @@ const customStringify = (value, indent = 2) => {
         const items = value.map(item => customStringify(item, indent + 2));
         return "[\n" + spacer + items.join(",\n" + spacer) + "\n" + ' '.repeat(indent - 2) + "]";
     }
-    const tshirtOrder = ["3xs", "2xs", "xs", "s", "md", "lg", "xl", "2xl", "3xl", "4xl", "5xl", "6xl", "7xl", "8xl", "9xl", "10xl", "11xl", "12xl", "13xl", "14xl", "15xl"];
+    const tshirtOrder = ["3xs", "2xs", "xs", "sm", "md", "lg", "xl", "2xl", "3xl", "4xl", "5xl", "6xl", "7xl", "8xl", "9xl", "10xl", "11xl", "12xl", "13xl", "14xl", "15xl"];
     const customComparator = (a, b) => {
         const indexA = tshirtOrder.indexOf(a);
         const indexB = tshirtOrder.indexOf(b);
